requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
docopt==0.6.2
flake8==6.0.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==5.12.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyjokes==0.6.0
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.15
yarg==0.1.9
